chore(dashboard): move formType to sub namespace
