+++ EMS/core-bundle/src/Resources/config/form.xml
@@ -319,7 +319,7 @@
             <argument type="service" id="ems.form.factories.objectChoiceListFactory" />
             <tag name="form.type"/>
         </service>
-        <service id="ems_core.form.dashboard_type" class="EMS\CoreBundle\Form\Form\DashboardType">
+        <service id="ems_core.form.dashboard_type" class="EMS\CoreBundle\Form\Form\Dashboard\DashboardType">
             <argument type="service" id="ems_core.dashboard.dashboards"/>
             <tag name="form.type"/>
         </service>
